+++ cartridge/webui/api-suggestions.lua
@@ -36,6 +36,16 @@ local force_apply_suggestion = gql_types.object({
     }
 })
 
+local disable_servers_suggestion = gql_types.object({
+    name = 'DisableServersSuggestion',
+    description =
+        'A suggestion to disable malfunctioning servers ' ..
+        ' in order to restore the quorum',
+    fields = {
+        uuid = gql_types.string.nonNull,
+    }
+})
+
 local function refine_uri()
     local topology_cfg = confapplier.get_readonly('topology')
     if topology_cfg == nil then
@@ -116,10 +126,35 @@ local function force_apply(_, _, info)
     return ret
 end
 
+local function disable_servers()
+    local topology_cfg = confapplier.get_readonly('topology')
+    if topology_cfg == nil then
+        return nil
+    end
+
+    local ret = {}
+
+    local refined_uri_list = topology.refine_servers_uri(topology_cfg)
+    for _, uuid, _ in fun.filter(topology.not_disabled, topology_cfg.servers) do
+        local member = membership.get_member(refined_uri_list[uuid])
+
+        if member.status ~= 'alive' and member.status ~= 'suspect' then
+            table.insert(ret, {uuid = uuid})
+        end
+    end
+
+    if next(ret) == nil then
+        return nil
+    end
+
+    return ret
+end
+
 local function get_suggestions(_, _, info)
     return {
         refine_uri = refine_uri(),
         force_apply = force_apply(nil, nil, info),
+        disable_servers = disable_servers()
     }
 end
 
@@ -134,6 +169,8 @@ local function init(graphql)
             fields = {
                 refine_uri = gql_types.list(refine_uri_suggestion.nonNull),
                 force_apply = gql_types.list(force_apply_suggestion.nonNull),
+                disable_servers =
+                    gql_types.list(disable_servers_suggestion.nonNull)
             }
         }),
         callback = module_name .. '.get_suggestions',